crud fro image

+++ app/Http/Controllers/PostsController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Post;
+
+use Illuminate\Support\Facades\Storage;
 use DB;
 
 class PostsController extends Controller
@@ -56,12 +58,34 @@ class PostsController extends Controller
     {
       $this->validate(request(),[
         'title'=>'required',
-        'body'=>'required'
+        'body'=>'required',
+        'cover_image'=>'image|nullable|max:1999'
       ]);
 
+      //handle file upload
+      if($request->hasFile('cover_image')){
+        //get filename with extension
+        $filenameWithExt=$request->file('cover_image')->getClientOriginalName();
+        //get just filename
+        $filename=pathinfo($filenameWithExt,PATHINFO_FILENAME);
+        //get just extension
+        $extension=$request->file('cover_image')->getClientOriginalExtension();
+        //filename to store
+        $fileNameToStore=$filename.'_'.time().'.'.$extension;
+        //upload image
+        $path=$request->file('cover_image')->storeAs('public/cover_images',$fileNameToStore);
+        
+      }else
+
+      {
+        $fileNameToStore='noimage.jpg';
+      }
+      
+           
         $post= new Post;
         $post->title= $request->title;
         $post->body= $request->body;
+        $post -> cover_image= $fileNameToStore;
         $post ->user_id= auth()->user()->id;
         $post->save();
 
@@ -109,12 +133,32 @@ class PostsController extends Controller
     {
       $this->validate(request(),[
         'title'=>'required',
-        'body'=>'required'
+        'body'=>'required',
+        'cover_image'=>'image|nullable|max:1999'
       ]);
+
+      //handle file upload
+      if($request->hasFile('cover_image')){
+        //get filename with extension
+        $filenameWithExt=$request->file('cover_image')->getClientOriginalName();
+        //get just filename
+        $filename=pathinfo($filenameWithExt,PATHINFO_FILENAME);
+        //get just extension
+        $extension=$request->file('cover_image')->getClientOriginalExtension();
+        //filename to store
+        $fileNameToStore=$filename.'_'.time().'.'.$extension;
+        //upload image
+        $path=$request->file('cover_image')->storeAs('public/cover_images',$fileNameToStore);
+        
+      }
+      
 
         $post =Post::find($id);
         $post->title= $request->title;
         $post->body= $request->body;
+        if($request->hasFile('cover_image')){
+          $post -> cover_image= $fileNameToStore;
+        }
         $post ->user_id= auth()->user()->id;
         $post->save();
 
@@ -129,8 +173,17 @@ class PostsController extends Controller
      */
     public function destroy($id)
     {
+
+
+      
     
       $post =Post::find($id);
+
+      //if img is not empty,delete it
+     if($post -> cover_image !== 'noimage.jpg'){
+      storage::delete('public/cover_images/'.$post->cover_image);
+      }
+
       $post->delete();
 
       if(auth()->user()->id !== $post->user_id){
